+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Feat: Support Attachment.addToTransactions (#709)
 * Fix: captureTransaction should return emptyId when transaction is discarded (#713)
 
 # 6.3.0-beta.3